cloud_function/requirements.txt
numpy
pandas
tensorflow==2.4.0
tensorflow_hub
pillow
google-api-python-client==1.12.8
google-cloud-storage==1.35.0
gunicorn
python-dotenv
